config_util/requirements.txt
PyYAML ~= 6.0
toml ~= 0.10.2
pydantic ~= 1.8.2
pytest ~= 6.2.5
